Updated model paths and renames (Post, Comment)
