+++ zimmerman/test/test_comment_model.py
@@ -3,7 +3,7 @@ from uuid import uuid4
 from datetime import datetime
 
 from zimmerman.main import db
-from zimmerman.main.model.user import Comments, User, Posts
+from zimmerman.main.model.main import Comment, User, Post
 from zimmerman.test.base import BaseTestCase
 
 
@@ -24,7 +24,7 @@ class TestCommentModel(BaseTestCase):
         db.session.commit()
 
         # Create test post
-        post = Posts(
+        post = Post(
             owner_id=user.id,
             creator_public_id=user.public_id,
             content="Test content",
@@ -36,14 +36,14 @@ class TestCommentModel(BaseTestCase):
         db.session.commit()
 
         # Create comment
-        comment = Comments(
+        comment = Comment(
             creator_public_id=user.public_id, on_post=post.id, content="Test comment"
         )
 
         db.session.add(comment)
         db.session.commit()
 
-        self.assertTrue(isinstance(comment, Comments))
+        self.assertTrue(isinstance(comment, Comment))
 
 
 if __name__ == "__main__":
